requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
google-generativeai
